Restructure developer page navigation on app-interbit-io
