+++ packages/app-interbit-io/src/containers/Developers/Overview.js
@@ -9,13 +9,10 @@ import {
   LinkBarSlack,
   Divider
 } from 'lib-react-interbit'
-import DeveloperNavigation from '../../components/DeveloperNavigation'
-import getInterbitServices from '../../redux/getInterbitServices'
 import urls from '../../constants/urls'
 import layout from '../../constants/layout'
 
 const mapStateToProps = state => ({
-  interbitServices: getInterbitServices(state),
   linkBarContent: state.content.linkBars,
   ...state.content.developers
 })
@@ -26,7 +23,7 @@ class DevelopersOverview extends Component {
     const { overview, linkBarContent } = this.props
     const colLayout = layout.colLayout.developers
 
-    const overviewContent = (
+    return (
       <div className="ibweb-page dev-overview">
         <Row>
           <Col {...colLayout}>
@@ -35,19 +32,6 @@ class DevelopersOverview extends Component {
           </Col>
         </Row>
 
-        {/* product roadmap sidebar example
-        <Row className="ibweb-dev-sidebar-list-row">
-          <Col {...colLayout}>
-            <ContentBarDefault {...overview.intro.contentBars[0]} />
-          </Col>
-          <Col lg={4} md={12} className="ibweb-dev-sidebar-col">
-            <SidebarList
-              title={overview.productRoadmap.title}
-              contents={overview.productRoadmap.items}
-            />
-          </Col>
-        </Row>
-    */}
         <Row className="app-characteristics">
           <Col {...colLayout}>
             <h2>{overview.appsBuilt.title}</h2>
@@ -109,8 +93,6 @@ class DevelopersOverview extends Component {
         </Row>
       </div>
     )
-
-    return <DeveloperNavigation {...this.props} component={overviewContent} />
   }
 }
