+++ packages/app-interbit-io/src/containers/Developers/Resources.js
@@ -3,14 +3,10 @@ import { connect } from 'react-redux'
 import { Row, Col } from 'react-bootstrap'
 import { Markdown, LinkBarSlack, Divider } from 'lib-react-interbit'
 
-import DeveloperNavigation from '../../components/DeveloperNavigation'
-import getInterbitServices from '../../redux/getInterbitServices'
 import urls from '../../constants/urls'
 import layout from '../../constants/layout'
 
 const mapStateToProps = state => ({
-  interbitServices: getInterbitServices(state),
-  linkBarContent: state.content.linkBars,
   ...state.content.developers
 })
 
@@ -20,7 +16,7 @@ class DevelopersResources extends Component {
     const { resources } = this.props
     const colLayout = layout.colLayout.developers
 
-    const resourcesContent = (
+    return (
       <div className="ibweb-page dev-resources">
         <Row>
           <Col {...colLayout}>
@@ -58,8 +54,6 @@ class DevelopersResources extends Component {
         </Row>
       </div>
     )
-
-    return <DeveloperNavigation {...this.props} component={resourcesContent} />
   }
 }
