+++ packages/app-interbit-io/src/containers/Developers/Support.js
@@ -3,13 +3,10 @@ import { connect } from 'react-redux'
 import { Row, Col } from 'react-bootstrap'
 import { Markdown, LinkBar, LinkBarSlack, Divider } from 'lib-react-interbit'
 
-import DeveloperNavigation from '../../components/DeveloperNavigation'
-import getInterbitServices from '../../redux/getInterbitServices'
 import urls from '../../constants/urls'
 import layout from '../../constants/layout'
 
 const mapStateToProps = state => ({
-  interbitServices: getInterbitServices(state),
   linkBarContent: state.content.linkBars,
   ...state.content.developers
 })
@@ -20,7 +17,7 @@ class DevelopersSupport extends Component {
     const { support, linkBarContent } = this.props
     const colLayout = layout.colLayout.developers
 
-    const supportContent = (
+    return (
       <div className="ibweb-page dev-support">
         <Row className="ibweb-mg-md">
           <Col {...colLayout}>
@@ -50,8 +47,6 @@ class DevelopersSupport extends Component {
         </Row>
       </div>
     )
-
-    return <DeveloperNavigation {...this.props} component={supportContent} />
   }
 }
