--- packages/app-interbit-io/src/content/developers/appBlueprints.js
@@ -1,36 +1,0 @@
-import chairmanmeow from '../../assets/chairmanmeow.jpg'
-
-export default {
-  title: 'App Blueprints',
-  intro: {
-    text: `Turkey andouille frankfurter, landjaeger ribeye sausage shank alcatra pork chop. Tongue porchetta meatloaf tenderloin. Pork alcatra t-bone biltong. Burgdoggen turkey short loin, kevin hamburger filet mignon buffalo tenderloin pancetta brisket drumstick bresaola leberkas meatloaf jowl. Spare ribs leberkas rump corned beef sirloin swine cupim pork chop doner short ribs pancetta. Corned beef tongue flank short ribs leberkas short loin, capicola andouille. Tongue turkey capicola venison hamburger bacon filet mignon pork buffalo bresaola brisket.`
-  },
-  section2: {
-    title: `Section 2 title`,
-    text: `Turkey andouille frankfurter, landjaeger ribeye sausage shank alcatra pork chop. Tongue porchetta meatloaf tenderloin. Pork alcatra t-bone biltong. Burgdoggen turkey short loin, kevin hamburger filet mignon buffalo tenderloin pancetta brisket drumstick bresaola leberkas meatloaf jowl. Spare ribs leberkas rump corned beef sirloin swine cupim pork chop doner short ribs pancetta. Corned beef tongue flank short ribs leberkas short loin, capicola andouille. Tongue turkey capicola venison hamburger bacon filet mignon pork buffalo bresaola brisket.`,
-    contentBars: [
-      {
-        title: `Title of tutorial or resource`,
-        content: `Buffalo bacon bresaola, sirloin tenderloin shankle fatback venison kevin alcatra tail.`,
-        image: chairmanmeow,
-        className: 'image-lg',
-        callToAction: {
-          to: '#',
-          type: 'button',
-          text: 'Tail chicken kielbasa'
-        }
-      }
-    ]
-  },
-  section3: {
-    linkBars: [
-      {
-        title: 'Title bacon',
-        image: chairmanmeow,
-        content:
-          'Brisket short ribs pork chop, shankle ground round buffalo pork belly cupim beef tenderloin beef ribs bacon venison.',
-        to: '#'
-      }
-    ]
-  }
-}
