--- packages/app-interbit-io/src/content/developers/appDirectory.js
@@ -1,3 +1,0 @@
-export default {
-  title: 'App Directory'
-}
