--- packages/app-interbit-io/src/content/developers/bestPractices.js
@@ -1,53 +1,0 @@
-import chairmanmeow from '../../assets/chairmanmeow.jpg'
-
-export default {
-  title: 'Best Practices',
-  intro: {
-    text: `Turkey andouille frankfurter, landjaeger ribeye sausage shank alcatra pork chop. Tongue porchetta meatloaf tenderloin. Pork alcatra t-bone biltong. Burgdoggen turkey short loin, kevin hamburger filet mignon buffalo tenderloin pancetta brisket drumstick bresaola leberkas meatloaf jowl. Spare ribs leberkas rump corned beef sirloin swine cupim pork chop doner short ribs pancetta. Corned beef tongue flank short ribs leberkas short loin, capicola andouille. Tongue turkey capicola venison hamburger bacon filet mignon pork buffalo bresaola brisket.`
-  },
-  section2: {
-    title: 'Section title',
-    text: `Chicken boudin pork chop cow flank brisket, tongue strip steak landjaeger. Bacon beef ribs flank, ribeye venison short loin leberkas shankle sirloin kevin strip steak prosciutto. Prosciutto flank cupim tenderloin capicola hamburger chuck.`,
-    contentBars: [
-      {
-        title: `Title of tutorial or resource`,
-        content: `Buffalo bacon bresaola, sirloin tenderloin shankle fatback venison kevin alcatra tail.`,
-        image: chairmanmeow,
-        className: 'image-sm',
-        callToAction: {
-          to: '#',
-          type: 'button',
-          text: 'Tail chicken kielbasa'
-        }
-      }
-    ]
-  },
-  section3: {
-    title: `Section 3 title`,
-    text: `Spicy jalapeno bacon ipsum dolor amet alcatra shankle tongue salami t-bone. Sausage pork loin pancetta, burgdoggen strip steak flank kevin. Leberkas drumstick swine, meatball bacon cupim meatloaf ball tip. Drumstick shankle short loin fatback pastrami porchetta. Meatball porchetta pancetta bresaola, hamburger burgdoggen sirloin flank bacon pork chop pastrami jerky short loin.`,
-    contentBars: [
-      {
-        title: `Title of tutorial or resource`,
-        content: `Buffalo bacon bresaola, sirloin tenderloin shankle fatback venison kevin alcatra tail.`,
-        image: chairmanmeow,
-        className: 'image-sm',
-        callToAction: {
-          to: '#',
-          type: 'button',
-          text: 'Tail chicken kielbasa'
-        }
-      }
-    ]
-  },
-  section4: {
-    linkBars: [
-      {
-        title: 'Title bacon',
-        image: chairmanmeow,
-        content:
-          'Brisket short ribs pork chop, shankle ground round buffalo pork belly cupim beef tenderloin beef ribs bacon venison.',
-        to: '#'
-      }
-    ]
-  }
-}
