--- packages/app-interbit-io/src/content/developers/building.js
@@ -1,23 +1,0 @@
-import dogesm from '../../assets/dogesm.jpg'
-
-export default {
-  title: 'Building Interbit Apps',
-  intro: {
-    text: `Bacon ipsum dolor amet drumstick pork ham hock cupim chuck capicola shoulder alcatra turkey biltong frankfurter brisket tongue ball tip shank. Kevin pig salami bacon, fatback short ribs doner jowl pastrami kielbasa. Capicola pork chicken picanha pork chop prosciutto pork loin short ribs sausage pork belly kielbasa pastrami doner meatball. Sausage tenderloin flank bresaola, turducken pork loin shank jowl. Ham frankfurter bresaola cow ham hock alcatra picanha sausage turkey buffalo beef ribs rump short loin.`
-  },
-  section2: {
-    title: 'Section title.',
-    text: `Alcatra prosciutto landjaeger venison, ham kielbasa drumstick. Burgdoggen shoulder andouille jowl kevin doner. T-bone prosciutto sirloin brisket sausage filet mignon ball tip biltong pork belly meatball landjaeger doner drumstick picanha. Pastrami landjaeger shoulder biltong, jowl fatback corned beef swine tri-tip tenderloin. Flank porchetta swine ball tip, meatball tenderloin tri-tip short loin. Sirloin salami venison jowl, bresaola ham burgdoggen tail ball tip chicken hamburger.`
-  },
-  section3: {
-    linkBars: [
-      {
-        title: 'Title bacon',
-        image: dogesm,
-        content:
-          'Brisket short ribs pork chop, shankle ground round buffalo pork belly cupim beef tenderloin beef ribs bacon venison.',
-        to: '#'
-      }
-    ]
-  }
-}
