--- packages/app-interbit-io/src/content/developers/identity.js
@@ -1,41 +1,0 @@
-import chairmanmeow from '../../assets/chairmanmeow.jpg'
-
-export default {
-  title: 'Identity Blueprint',
-  intro: {
-    title: 'Blueprint title introducing why service needs to exist',
-    text:
-      'Kevin sirloin pancetta, short loin filet mignon landjaeger porchetta ball tip biltong bresaola venison buffalo meatball. Burgdoggen strip steak shankle t-bone tongue porchetta pork loin ground round hamburger fatback. Ribeye corned beef prosciutto beef ribs chuck boudin kielbasa. Boudin pork brisket picanha doner hamburger beef ribs pork belly burgdoggen venison frankfurter. Pork chop doner beef ribs, rump pork belly tongue pig pork.'
-  },
-  section2: {
-    title: 'Section 2 title',
-    text:
-      'Frankfurter tongue bresaola brisket, chuck pork belly spare ribs ham beef ribs tail strip steak meatball hamburger fatback doner. Pancetta jerky shankle turkey ball tip ham hock spare ribs hamburger ground round boudin burgdoggen landjaeger tenderloin. Kevin cupim pork, jowl pork belly venison strip steak bresaola andouille flank pastrami turkey ground round chuck. Pork chop venison pastrami brisket boudin shankle tail pork loin chuck jowl ham pork chicken.'
-  },
-  quote: {
-    content:
-      'Shank chuck pork belly shankle strip steak short loin meatball ribeye swine corned beef t-bone ball tip sirloin. Strip steak salami shankle turkey pork.',
-    author: 'Jane Doe',
-    publication: 'Publication name',
-    image: chairmanmeow
-  },
-  section3: {
-    cards: [
-      {
-        title: 'A meatier lorem ipsum generator',
-        content:
-          'Burgdoggen prosciutto jerky, ground round t-bone salami rump. Sausage meatball frankfurter jerky bacon buffalo pork kielbasa brisket cow pastrami ham.',
-        image: chairmanmeow
-      }
-    ]
-  },
-  code: {
-    markdown: `
-    describe('Component', () => {
-      let node
-    
-      beforeEach(() => {
-        node = document.createElement('div')
-        `
-  }
-}
