+++ packages/app-interbit-io/src/content/developers/index.js
@@ -46,7 +46,7 @@ export default {
       }
     }
   ],
-  sideBar: [
+  sidebar: [
     {
       title: 'Start Here',
       items: [
