--- packages/app-interbit-io/src/content/developers/projects.js
@@ -1,3 +1,0 @@
-export default {
-  title: 'Projects'
-}
